+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Fix monitor not detected in Wayland.
 * Fix window restore by command not returning to same position.
 * Fix `WindowVars::monitor` changes not moving window to new monitor.
     - Added demonstration in `window` example.